[fix] Screen name 수정
