+++ app/src/main/res/values/strings.xml
@@ -39,5 +39,6 @@
     <string name="report_team_percent">팀 인증율</string>
     <string name="report_my_percent">내 인증율</string>
     <string name="report_early_bird">이번 회차 얼리버드</string>
+    <string name="report_count">%d회차 리포트</string>
 
 </resources>
